+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * ifdef preprocessor, uses [gpp](https://github.com/Kestrer/gpp)
 * added abstract console
 * removed old fetchresponseprovider and added new Fetch api (wip)
+* added greco://cache as a simple caching module
  
 0.0.1